third_party/WebKit/LayoutTests/inspector/console/console-format-es6-expected.txt

CONSOLE MESSAGE: line 11: [object Object]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 11: Symbol()
CONSOLE MESSAGE: line 12: Symbol()
CONSOLE MESSAGE: line 11: Symbol(a)
CONSOLE MESSAGE: line 12: Symbol(a)
CONSOLE MESSAGE: line 11: [object Object]
CONSOLE MESSAGE: line 12: [object Object]
Tests that console properly displays information about ES6 features.

Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0} console-format-es6.html:11
[Promise]0: Promiselength: 1__proto__: Array[0] console-format-es6.html:12
globals[0]
Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
Symbol() console-format-es6.html:11
[Symbol()]0: Symbol()length: 1__proto__: Array[0] console-format-es6.html:12
globals[1]
Symbol()
Symbol(a) console-format-es6.html:11
[Symbol(a)]0: Symbol(a)length: 1__proto__: Array[0] console-format-es6.html:12
globals[2]
Symbol(a)
Object {a: Symbol(), Symbol(): 1, Symbol(a): 2}a: Symbol()Symbol(): 1Symbol(a): 2__proto__: Object console-format-es6.html:11
[Object]0: Objectlength: 1__proto__: Array[0] console-format-es6.html:12
globals[3]
Object {a: Symbol(), Symbol(): 1, Symbol(a): 2}a: Symbol()Symbol(): 1Symbol(a): 2__proto__: Object
